+++ CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Please also read the [Upgrade Guide](https://github.com/katzer/cordova-plugin-local-notifications/wiki/Upgrade-Guide) for more information.
 
+#### Unreleased
+- Fix required for Android 13 regarding notification permissions (https://outsystemsrd.atlassian.net/browse/RMET-1833)
+
 #### Version 0.9.8 (04.11.2021)
 - New plugin release to include metadata tag in Extensibility Configurations in the OS wrapper